chore: move lookupMap util to virtual-mod
