+++ packages/astro/src/content/vite-plugin-content-virtual-mod.ts
@@ -2,14 +2,20 @@ import fsMod from 'node:fs';
 import type { Plugin } from 'vite';
 import type { AstroSettings } from '../@types/astro.js';
 import { VIRTUAL_MODULE_ID } from './consts.js';
+import glob, { type Options as FastGlobOptions } from 'fast-glob';
 import {
 	getContentEntryConfigByExtMap,
 	getDataEntryExts,
 	getContentPaths,
 	getExtGlob,
-	getStringifiedLookupMap,
+	getEntryCollectionName,
+	getContentEntryIdAndSlug,
+	getEntrySlug,
+	getDataEntryId,
 } from './utils.js';
 import { rootRelativePath } from '../core/util.js';
+import { extname } from 'node:path';
+import { fileURLToPath, pathToFileURL } from 'node:url';
 
 interface AstroContentVirtualModPluginParams {
 	settings: AstroSettings;
@@ -68,4 +74,86 @@ export function astroContentVirtualModPlugin({
 			}
 		},
 	};
+}
+
+export type ContentLookupMap = {
+	[collectionName: string]: { type: 'content' | 'data'; entries: { [lookupId: string]: string } };
+};
+
+export async function getStringifiedLookupMap({
+	contentDir,
+	contentEntryConfigByExt,
+	dataEntryExts,
+	root,
+	fs,
+}: {
+	contentEntryConfigByExt: ReturnType<typeof getContentEntryConfigByExtMap>;
+	dataEntryExts: string[];
+	contentDir: URL;
+	root: URL;
+	fs: typeof fsMod;
+}) {
+	const globOpts: FastGlobOptions = {
+		absolute: true,
+		cwd: fileURLToPath(root),
+		fs: {
+			readdir: fs.readdir.bind(fs),
+			readdirSync: fs.readdirSync.bind(fs),
+		},
+	};
+
+	let lookupMap: ContentLookupMap = {};
+	const relContentDir = rootRelativePath(root, contentDir, false);
+	const contentGlob = await glob(
+		`${relContentDir}**/*${getExtGlob([...contentEntryConfigByExt.keys()])}`,
+		globOpts
+	);
+
+	await Promise.all(
+		contentGlob.map(async (filePath) => {
+			const contentEntryType = contentEntryConfigByExt.get(extname(filePath));
+			if (!contentEntryType) return;
+			const collection = getEntryCollectionName({ contentDir, entry: pathToFileURL(filePath) });
+			// Globbed entry outside of a collection
+			// Logs warning on type generation, safe to ignore in lookup map
+			if (!collection) return;
+
+			const { id, slug: generatedSlug } = await getContentEntryIdAndSlug({
+				entry: pathToFileURL(filePath),
+				contentDir,
+				collection,
+			});
+			lookupMap[collection] ??= { type: 'content', entries: {} };
+			const slug = await getEntrySlug({
+				id,
+				collection,
+				generatedSlug,
+				fs,
+				fileUrl: pathToFileURL(filePath),
+				contentEntryType,
+			});
+			lookupMap[collection].entries[slug] = rootRelativePath(root, filePath);
+		})
+	);
+
+	const dataGlob = await glob(`${relContentDir}**/*${getExtGlob(dataEntryExts)}`, globOpts);
+	await Promise.all(
+		dataGlob.map(async (filePath) => {
+			const collection = getEntryCollectionName({
+				contentDir,
+				entry: pathToFileURL(filePath),
+			});
+			if (!collection) return;
+
+			const id = getDataEntryId({
+				entry: pathToFileURL(filePath),
+				contentDir,
+				collection,
+			});
+			lookupMap[collection] ??= { type: 'data', entries: {} };
+			lookupMap[collection].entries[id] = rootRelativePath(root, filePath);
+		})
+	);
+
+	return JSON.stringify(lookupMap);
 }
